Create rest directory to hold the girder and publish clients
